Use long, not int in do_exhaustive (problem on 64bit machines).

+++ src/basemath/bibli1.c
@@ -1741,7 +1741,7 @@ static GEN
 do_exhaustive(GEN P, GEN N, int x, GEN B) 
 {
   GEN tst, sol = cget1(2*x + 2, t_VECSMALL);
-  int j, l;
+  long j, l;
 
   for (j = -x; j <= x; j++)
   {
